pyproject.toml
[project]
name = "faithful"
version = "3.0.0"
description = "A Discord chatbot that emulates the tone and typing style of provided example messages."
requires-python = ">=3.10"
dependencies = [
    "discord.py>=2.3",
    "tomli>=2.0; python_version < '3.11'",
    "duckduckgo-search>=7.0",
    "aiohttp>=3.9",
    "beautifulsoup4>=4.12",
]

[project.optional-dependencies]
openai = ["openai>=1.68"]
gemini = ["google-genai>=1.0"]
anthropic = ["anthropic>=0.40"]
all = [
    "openai>=1.68",
    "google-genai>=1.0",
    "anthropic>=0.40",
]
dev = [
    "openai>=1.68",
    "google-genai>=1.0",
    "anthropic>=0.40",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
faithful = "faithful.__main__:main"

[tool.setuptools.packages.find]
include = ["faithful*"]
